Made Update Core Warning Templatable
The update warning on the updates page now checks for a template file
in the theme folder. If present it uses the content here and if not
present it uses content in a default file in the inc folder. The
dashboard widget file was also moved into a inc folder in the plugin
root to keep files organised more.