Built Dashboard Content for new Dashboard Home
Added alls to a remote file to get content which can be changed for all
sites which use pxlcore and also called a dashboard-content file in the
theme is present so that custom dashboard content be be added for each
site. TODO - the call the external file will need to be the correct
location once decided. Currently looks for a local file.